retoques de errores leves solucionados y test añadidos de 0.5 finalizados (ACABADO)

+++ cws_services_jpa/src/uo/ri/cws/domain/SparePart.java
@@ -77,7 +77,7 @@ public class SparePart extends BaseEntity
 		return price;
 	}
 
-	public double getStock() {
+	public int getStock() {
 		return stock;
 	}
